+++ Sources/ParseSwift/Objects/ParseCloudUser.swift
@@ -15,4 +15,9 @@ import Foundation
 public protocol ParseCloudUser: ParseUser {
     /// The session token of the `ParseUser`.
     var sessionToken: String? { get set }
+    /// The number of unsuccessful login attempts.
+    var _failed_login_count: Int? { get }
+    /// The date the lockout expires. After this date, the `ParseUser`
+    /// can attempt to login again.
+    var _account_lockout_expires_at: Date? { get }
 }